# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flower Shop API",
        "description": "API documentation for the Flower Shop system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lấy danh sách sản phẩm",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tạo sản phẩm mới với nhiều size và công thức hoa",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "status",
                                    "sizes[0][size]",
                                    "sizes[0][recipes][0][flower_id]",
                                    "sizes[0][recipes][0][quantity]"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Tên sản phẩm",
                                        "type": "string",
                                        "example": "Bó hoa cưới đẹp"
                                    },
                                    "description": {
                                        "description": "Mô tả sản phẩm",
                                        "type": "string",
                                        "example": "Mẫu bó hoa cưới với nhiều hoa baby"
                                    },
                                    "category_id": {
                                        "description": "ID danh mục",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "description": "Trạng thái sản phẩm",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "Ảnh sản phẩm",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "sizes[0][size]": {
                                        "description": "Tên size 1",
                                        "type": "string",
                                        "example": "Nhỏ"
                                    },
                                    "sizes[0][recipes][0][flower_id]": {
                                        "description": "ID hoa thứ 1 cho size Nhỏ",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sizes[0][recipes][0][quantity]": {
                                        "description": "Số lượng hoa thứ 1 cho size Nhỏ",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "sizes[0][recipes][1][flower_id]": {
                                        "description": "ID hoa thứ 2 cho size Nhỏ",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "sizes[0][recipes][1][quantity]": {
                                        "description": "Số lượng hoa thứ 2 cho size Nhỏ",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "sizes[1][size]": {
                                        "description": "Tên size 2",
                                        "type": "string",
                                        "example": "Lớn"
                                    },
                                    "sizes[1][recipes][0][flower_id]": {
                                        "description": "ID hoa cho size Lớn",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sizes[1][recipes][0][quantity]": {
                                        "description": "Số lượng hoa cho size Lớn",
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Tìm kiếm sản phẩm theo tên sản phẩm, tên danh mục, hoặc tên hoa (ưu tiên theo thứ tự này)",
                "operationId": "33867295f3d0876f2cb47a1e6da11e02",
                "parameters": [
                    {
                        "name": "product",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm (tên sản phẩm, tên danh mục hoặc tên hoa)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tìm theo tên sản phẩm (nếu muốn lọc riêng)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Tìm theo ID danh mục (nếu muốn lọc riêng)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "flower_name",
                        "in": "query",
                        "description": "Tìm theo tên hoa (nếu muốn lọc riêng)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm tìm kiếm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lấy thông tin sản phẩm theo ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cập nhật sản phẩm với nhiều size và công thức hoa",
                "operationId": "b4b789a4a3d0c47897f43991a5f12372",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID sản phẩm cần cập nhật",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name",
                                    "category_id",
                                    "status",
                                    "sizes[0][size]",
                                    "sizes[0][recipes][0][flower_id]",
                                    "sizes[0][recipes][0][quantity]"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Phương thức HTTP (đặt là PUT để cập nhật)",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Bó hoa cưới đỏ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bó hoa cưới rực rỡ"
                                    },
                                    "image": {
                                        "description": "Ảnh sản phẩm (tùy chọn)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "status": {
                                        "description": "Trạng thái sản phẩm",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "description": "ID danh mục sản phẩm",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sizes[0][size]": {
                                        "type": "string",
                                        "example": "Nhỏ"
                                    },
                                    "sizes[0][recipes][0][flower_id]": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sizes[0][recipes][0][quantity]": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "sizes[0][recipes][1][flower_id]": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "sizes[0][recipes][1][quantity]": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "sizes[1][size]": {
                                        "type": "string",
                                        "example": "Lớn"
                                    },
                                    "sizes[1][recipes][0][flower_id]": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sizes[1][recipes][0][quantity]": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật sản phẩm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Xóa sản phẩm",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/products/category/{categoryId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lấy danh sách sản phẩm theo danh mục",
                "operationId": "d9a719d13ea154ac76fb531fe1b679dc",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách sản phẩm theo danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Danh mục không tồn tại"
                    }
                }
            }
        },
        "/api/products/{id}/hide": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Ẩn sản phẩm",
                "operationId": "b352e3b23ebdc0ae9ec12abe989d0b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product hidden successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
        "/api/products/{id}/stock": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Kiểm tra tồn kho của sản phẩm theo ID",
                "operationId": "fcd232ac145eef6eac3b0ba58d3f7c78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tồn kho của sản phẩm",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 9
                    },
                    "name": {
                        "type": "string",
                        "example": "Hoa hồng"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nhập lô hoa hồng"
                    },
                    "status": {
                        "type": "boolean",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/image.jpg"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeDetailResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token JWT vào ô Authorization",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}